+++ packages/ui-components/src/lib/index.ts
@@ -117,7 +117,12 @@ export { lightCodeMirrorTheme, darkCodeMirrorTheme } from './utils/codeMirrorThe
 
 // Stores
 export { default as transactionStore } from './stores/transactionStore';
-export { cachedWritableStore } from './storesGeneric/cachedWritableStore';
+export {
+	cachedWritableStore,
+	cachedWritableIntOptional,
+	cachedWritableStringOptional,
+	cachedWritableString
+} from './storesGeneric/cachedWritableStore';
 
 // Assets
 export { default as logoLight } from './assets/logo-light.svg';
